Add hmm_document to fixtures
