+++ tests/api/test_hmm.py
@@ -1,54 +1,4 @@
-import pytest
 from aiohttp.test_utils import make_mocked_coro
-
-
-@pytest.fixture
-def hmm_document():
-    return {
-        "_id": "f8666902",
-        "count": 4,
-        "length": 199,
-        "names": [
-            "ORF-63",
-            "ORF67",
-            "hypothetical protein"
-        ],
-        "entries": [
-            {
-                "gi": "438000415",
-                "organism": "Thysanoplusia orichalcea nucleopolyhedrovirus",
-                "name": "hypothetical protein",
-                "accession": "YP_007250520.1"
-            },
-            {
-                "gi": "114679914",
-                "organism": "Leucania separata nucleopolyhedrovirus",
-                "name": "ORF67",
-                "accession": "YP_758364.1"
-            },
-            {
-                "gi": "209170953",
-                "organism": "Agrotis ipsilon multiple nucleopolyhedrovirus",
-                "name": "agip69",
-                "accession": "YP_002268099.1"
-            },
-            {
-                "gi": "90592780",
-                "organism": "Agrotis segetum nucleopolyhedrovirus",
-                "name": "ORF-63",
-                "accession": "YP_529733.1"
-            }
-        ],
-        "total_entropy": 101.49,
-        "families": {
-            "Baculoviridae": 3
-        },
-        "genera": {
-            "Alphabaculovirus": 3
-        },
-        "cluster": 3463,
-        "mean_entropy": 0.51
-    }
 
 
 async def test_find(mocker, spawn_client, hmm_document):
